Added 2 CRE spacing, orientation, and distance library design
Never added note for this script.

This is the base script for library generation.

Uses a lot of Kim's definitions for checking the library, reading in and writing and some more specific ones such as scrambling a sequence.

First designs consist of scrambling 10 bp of the new backgrounds in 5 bp increments.

Second design places 0-6 consensus CREs with flanks in equally-spaced positions along the new backgrounds.

Design added is still in the works. Takes either the consensus with flanks or its reverse compliment and varies that across two sites at different spacing from one another. Both CREs and spacing are then moved across the old set of 3 backgrounds.